requirements.txt
# Ring Image Search System - Requirements
# Install with: pip install -r requirements.txt

tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
flask>=2.0.0
flask-cors>=3.0.0
tqdm>=4.64.0
